+++ sources/Google.Solutions.IapDesktop.Extensions.Os.Test/Google.Solutions.IapDesktop.Extensions.Os.Test.csproj
@@ -114,10 +114,10 @@
     <Compile Include="Inventory\TestWuaPackage.cs" />
     <Compile Include="Inventory\TestWuaPackageType.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Services\Windows\TestDomainJoinService.cs" />
-    <Compile Include="Services\Windows\TestDomainJoinServiceStartupScript.cs" />
-    <Compile Include="Services\Windows\TestNetworkCredentialExtensions.cs" />
-    <Compile Include="Services\Windows\TestStartupScriptOperation.cs" />
+    <Compile Include="Services\ActiveDirectory\TestDomainJoinService.cs" />
+    <Compile Include="Services\ActiveDirectory\TestDomainJoinServiceStartupScript.cs" />
+    <Compile Include="Services\ActiveDirectory\TestNetworkCredentialExtensions.cs" />
+    <Compile Include="Services\ActiveDirectory\TestStartupScriptOperation.cs" />
     <Compile Include="Views\InstanceProperties\TestInstancePropertiesInspectorModel.cs" />
     <Compile Include="Views\InstanceProperties\TestInstancePropertiesInspectorViewModel.cs" />
     <Compile Include="Services\Inventory\TestInventoryService.cs" />
