+++ sources/Google.Solutions.IapDesktop.Extensions.Os/Services/ActiveDirectory/DomainJoinService.StartupScript.ps1
