+++ sources/Google.Solutions.IapDesktop.Extensions.Os/Services/ActiveDirectory/StartupScriptOperation.cs
@@ -33,7 +33,7 @@ using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Google.Solutions.IapDesktop.Extensions.Os.Services.Windows
+namespace Google.Solutions.IapDesktop.Extensions.Os.Services.ActiveDirectory
 {
     /// <summary>
     /// A long-running operation that involves swapping out
